+++ boards/arm64/a64/pinephone/src/pinephone_lcd.h
@@ -1,0 +1,1 @@
+int pinephone_panel_init(void);
